pyproject.toml
[tool.poetry]
name = "daggerml-cli"
version = "0.1.0"
description = ""
authors = ["Micha Niskin <micha.niskin@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
daggerml = "daggerml_cli.main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.4.0"}
requests = "^2.27.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
